Prometheus collector for machine metrics
