+++ metrics/testdata/prometheus_machine_metrics
@@ -1,0 +1,29 @@
+# HELP cadvisor_version_info A metric with a constant '1' value labeled by kernel version, OS version, docker version, cadvisor version & cadvisor revision.
+# TYPE cadvisor_version_info gauge
+cadvisor_version_info{cadvisorRevision="abcdef",cadvisorVersion="0.16.0",dockerVersion="1.8.1",kernelVersion="4.1.6-200.fc22.x86_64",osVersion="Fedora 22 (Twenty Two)"} 1
+# HELP machine_cpu_logical_cores Number of logical CPU cores
+# TYPE machine_cpu_logical_cores gauge
+machine_cpu_logical_cores{boot_id="",machine_id="",system_uuid=""} 4
+# HELP machine_cpu_physical_cores Number of physical CPU cores
+# TYPE machine_cpu_physical_cores gauge
+machine_cpu_physical_cores{boot_id="",machine_id="",system_uuid=""} 1
+# HELP machine_cpu_sockets Number of CPU sockets
+# TYPE machine_cpu_sockets gauge
+machine_cpu_sockets{boot_id="",machine_id="",system_uuid=""} 1
+# HELP machine_dimm_capacity_bytes Total RAM DIMM capacity (all types memory modules) value labeled by dimm type
+# TYPE machine_dimm_capacity_bytes gauge
+machine_dimm_capacity_bytes{boot_id="",machine_id="",system_uuid="",type="Non-volatile-RAM"} 2.168421613568e+12
+machine_dimm_capacity_bytes{boot_id="",machine_id="",system_uuid="",type="Unbuffered-DDR4"} 4.12316860416e+11
+machine_dimm_capacity_bytes{boot_id="",machine_id="",system_uuid="",type="all"} 2.580738473984e+12
+# HELP machine_dimm_count Number of RAM DIMM (all types memory modules) value labeled by dimm type
+# TYPE machine_dimm_count gauge
+machine_dimm_count{boot_id="",machine_id="",system_uuid="",type="Non-volatile-RAM"} 8
+machine_dimm_count{boot_id="",machine_id="",system_uuid="",type="Unbuffered-DDR4"} 12
+machine_dimm_count{boot_id="",machine_id="",system_uuid="",type="all"} 20
+# HELP machine_nvm_capacity NVM capacity value labeled by NVM mode (memory mode or app direct mode)
+# TYPE machine_nvm_capacity gauge
+machine_nvm_capacity{boot_id="",machine_id="",mode="app_direct_mode",system_uuid=""} -1
+machine_nvm_capacity{boot_id="",machine_id="",mode="memory_mode",system_uuid=""} -1
+# HELP machine_scrape_error 1 if there was an error while getting machine metrics, 0 otherwise
+# TYPE machine_scrape_error gauge
+machine_scrape_error 0
