+++ metrics/testdata/prometheus_machine_metrics_failure
@@ -1,0 +1,3 @@
+# HELP machine_scrape_error 1 if there was an error while getting machine metrics, 0 otherwise
+# TYPE machine_scrape_error gauge
+machine_scrape_error 1
